requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
comm==0.1.2
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
Flask==2.2.2
Flask-HTTPAuth==4.7.0
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.2
idna==3.4
importlib-metadata==5.1.0
ipykernel==6.19.2
ipython==8.7.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter-client==7.4.8
jupyter-core==5.1.0
kiwisolver==1.4.4
line-bot-sdk==2.3.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
multidict==6.0.3
nest-asyncio==1.5.6
numpy==1.23.5
packaging==22.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pyzmq==24.0.1
requests==2.28.1
scikit-learn==1.2.0
scipy==1.9.3
six==1.16.0
stack-data==0.6.2
TA-Lib==0.4.25
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.7.1
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.11.0
